+++ ec/types/adv.py
@@ -1,3 +1,8 @@
+"""
+adv
+===
+Advanced types.
+"""
 import shlex
 
 from ..modules.classes import CustomType, HandledException
@@ -5,6 +10,9 @@ from ..modules.helpers import list2dict
 
 class t2t(CustomType):
   """Convert a ec task into a type.
+  
+  Args:
+    __ec__task__: Any ec task.
   """
   def __init__(self, __ec__task__, **InArgs):
     __ec_member__ = __ec__task__.__ec_member__
@@ -22,19 +30,19 @@ class t2t(CustomType):
     return self.Task.__collect_n_call__(**Args)
     
 class chain(CustomType):
+  """Combines mutiple types into one.
+  
+  Args:
+    *Types (Type): The types to chain.
+    
+  Kwargs:
+    desc (str): The description for the chain.
+    
+  Example:
+  
+    @arg(type=chain(exists, isabs), desc="an existing abs path")
+  """
   def __init__(self, *Types, **Kwargs):
-    """Combines mutiple types into one.
-    
-    Args:
-      *Types (Type): The types to chain.
-      
-    Kwargs:
-      desc (str): The description for the chain.
-      
-    Example:
-    
-      @arg(type=chain(exists, isabs), desc="an existing abs path")
-    """
     CustomType.__init__(self, Kwargs.get('desc'))
     
     self.Types = Types
@@ -50,3 +58,33 @@ class chain(CustomType):
     
   def __str__(self):
     return getattr(self, 'desc', str(self.CurrentType) if self.CurrentType else '')
+    
+class invert(CustomType):
+  """Inverts the given type.
+  
+  ie: Only failed values are qualified.
+  
+  Args:
+    desc (str): The description for the type.
+    
+  Example:
+  
+    @arg(type=invert(exists), desc="a free path")
+  """
+  def __init__(self, Type, desc=None):
+    CustomType.__init__(self, desc)
+    
+    self.Type = Type
+    
+  def __call__(self, val):
+    
+    try:
+      self.Type(val)
+      
+    except (ValueError, TypeError):
+      return val
+    
+    raise ValueError()
+    
+  def __str__(self):
+    return getattr(self, 'desc', 'a value that is not %s' % str(self.Type))